Update: FSS Extended List should be non-recursive.
Update specification to make this more clear.

The Extended List read functions are updated to handle this.
The Extended List read program is essentially copied from the Basic List read program and then updated accordingly.
The Extended List write functions and programs need to not escape the opening brace inside a content.